Fixed timestamps, command console

+++ home.py
@@ -4,6 +4,7 @@ from rpi_lcd import LCD
 from signal import signal, SIGTERM, SIGHUP, pause
 from threading import Thread
 import time
+from pathlib import Path
 
 class Arbiter():
 	def __init__(self):
@@ -11,19 +12,21 @@ class Arbiter():
 		self.__pir_pin=11
 		self.__dht_pin=12
 		self.__lcd = None
-		#self.__lcd = LCD()
+		self.__lcd = LCD()
 		self.__temperature = 0
 		self.__humidity = 0
 		self.__running = True
 		self.__DHT_SENSOR=DHT.DHT11
-		self.__currTime = time.localtime()
-		self.__filename = './logs/sensor_data_'+str(time.localtime()[3])+str(time.localtime()[4])+str(time.localtime()[5])+'.csv'
-		with open(self.__filename, 'w', encoding="utf-8") as file:
-			file.write('time,humidity,temperature')
-		del self.__currTime
 		self.__threads = []
 		signal(SIGTERM, self.safe_exit)
 		signal(SIGHUP, self.safe_exit)
+
+		self.__currTime = time.localtime()
+		self.__filename = './logs/sensor_data_'+str(self.__currTime[0])+str(self.__currTime[1])+str(self.__currTime[2])+'.csv'
+		if Path('~./logs').is_dir():
+			if not Path('~'+self.__filename).is_file():
+				with open(self.__filename, 'w', encoding="utf-8") as file:
+					file.write('time,humidity,temperature')
 		
 	def __del__(self):
 		self.clear()
@@ -38,7 +41,7 @@ class Arbiter():
 
 	def clear(self):
 		GPIO.cleanup()
-		#self.__lcd.clear()
+		self.__lcd.clear()
 
 	def safe_exit(signum, frame):
 		exit(1)
@@ -49,10 +52,19 @@ class Arbiter():
 		self.__threads.append(Thread(target=self.lcd, name='lcd'))
 		self.__threads[len(self.__threads)-1].start()
 		exit_list = ['quit', 'exit', 'q']
+		temperature = ['temp', 'temperature', 'tp', 'humd', 'humidity', 'hd']
 		while self.__running:
 			user_input = input('Command: ')
 			if user_input.lower() in exit_list:
 				self.__running = False
+			elif user_input.lower() in temperature:
+				print(time.ctime(), '- Temperature:{0:0.1f}C Humidity:{1:0.1f}%'.format(self.__temperature, self.__humidity))
+			elif user_input.lower() == 'help':
+				print('Currently supported commands are:\n')
+				[print(i) for i in exit_list]
+				[print(j) for j in temperature]
+			else:
+				print('Command not recognized.')
 		self.__running = False
 		self.__threads[len(self.__threads)-2].join()
 		self.__threads[len(self.__threads)-1].join()
@@ -68,8 +80,11 @@ class Arbiter():
 			motion = GPIO.input(self.__pir_pin)
 			if motion:
 				GPIO.output(self.__led_pin,True)
-				self.__lcd.text(str(time.localtime()[3])+':'+str(time.localtime()[4])+':'+str(time.localtime()[5]), 1)
+				currTime = time.localtime()
+				# self.__lcd.text(str(time.localtime()[3])+':'+str(time.localtime()[4])+':'+str(time.localtime()[5]), 1)
+				self.__lcd.texT(str(currTime[3])+':'+str(currTime[4])+':'+str(currTime[5]), 1)
 				self.__lcd.text("Tp:{0:0.1f}C Hd:{1:0.1f}%".format(self.__temperature, self.__humidity), 2)
+				del currTime
 				time.sleep(0.25)
 			else:
 				GPIO.output(self.__led_pin,False)
@@ -81,14 +96,22 @@ class Arbiter():
 			humidity, temperature = DHT.read(self.__DHT_SENSOR, self.__dht_pin)
 			if humidity is not None and temperature is not None:
 				self.__humidity, self.__temperature = humidity, temperature
-				with open(self.__filename, 'a', encoding="utf-8") as file:
-					file.write(str(time.time())+','+str(self.__humidity)+','+str(self.__temperature))
+				currTime = time.localtime()
+				if currTime[2] == self.__currTime[2]: # Use the current file
+					with open(self.__filename, 'a', encoding="utf-8") as file:
+						file.write(str(time.time())+','+str(self.__humidity)+','+str(self.__temperature))
+				else: # Create a new file
+					self.__filename = './logs/sensor_data_'+str(currTime[0])+str(currTime[1])+str(currTime[2])+'.csv'
+					with open(self.__filename, 'w', encoding="utf-8") as file:
+						file.write('time,humidity,temperature')
+						file.write(str(time.time())+','+str(self.__humidity)+','+str(self.__temperature))
+				del currTime
 			else:
-				print("Could not read data from humidity sensor.")
+				print(time.ctime(), "- Could not read data from humidity sensor.")
 			time.sleep(1)
 # end Arbiter
 
 if __name__ == "__main__":
-	print(str(time.localtime()[3])+':'+str(time.localtime()[4])+':'+str(time.localtime()[5]))
+	print(time.ctime())
 	newArbiter = Arbiter()
 	newArbiter.start()
